// src/app/core/models/annotation.ts
export interface Annotation {
    id: string;
    versionId: string;
    type: string;
    data: string;
    userId: string;
    sessionId: string;
    color: string;
    strokeW: number;
    strokeStyle: string;
    x: number;
    y: number;
    title: string;
    text: string;
    createdAt: string;
    updatedAt: string;
}